update gps code

+++ Project1/3년차/gps/gps_with_gyro2.cpp
@@ -355,9 +355,9 @@ double get_gpsData(char *chrBuf, int num)
     }
 }
 //0x58 Ground Speed data 추출
-float get_groundSpeed(char *chrBuf, int num)
+double get_groundSpeed(char *chrBuf, int num)
 {
-    float gh; float gy; float gv;
+    double gh; double gy; double gv;
     unsigned int tmp[8];
     unsigned char i;
 
@@ -366,15 +366,15 @@ float get_groundSpeed(char *chrBuf, int num)
     }
 
     if (num==1){
-        gh = ((float)((tmp[1]<<8)|tmp[0]))/10;
+        gh = ((double)((tmp[1]<<8)|tmp[0]))/10;
         return gh;
     }
     else if (num==2){
-        gy = ((float)((tmp[3]<<8)|tmp[2]))/10;
+        gy = ((double)((tmp[3]<<8)|tmp[2]))/10;
         return gy;
     }
     else if (num==3){
-        gv = ((float)((tmp[7]<<24)|(tmp[6]<<16)|(tmp[5]<<8)|tmp[4]))/1000;
+        gv = ((double)((tmp[7]<<24)|(tmp[6]<<16)|(tmp[5]<<8)|tmp[4]))/1000;
         return gv;
     }
     else{
@@ -453,7 +453,7 @@ float roll; float pitch; float yaw;//0x53
 float mx; float my; float mz;//0x54
 float press; float h; //0x56
 double lon; double lat; double lon_dd; double lat_dd; double lon_mm; double lat_mm;//0x57
-float gh; float gy; float gv;//0x58
+double gh; double gy; double gv;//0x58
 float q0; float q1; float q2; float q3;//0x59
 float sn; float pdop; float hdop; float vdop;//0x5a
